Update Enemy AI
- Fix Follow, Detective BT

+++ Assets/Scripts/Controller/Animation/AnimationController.cs
@@ -20,7 +20,7 @@ namespace Controller.Animation
             _animator.SetFloat(AnimatorHash.AnimationSpeed, context.AnimationMultiplier);
             
             _animator.SetInteger(AnimatorHash.Weapon, (int)context.WeaponType);
-            _animator.SetInteger(AnimatorHash.Action, (int)context.ActionNumber);
+            _animator.SetInteger(AnimatorHash.Action, context.ActionNumber);
             
             // 샘플 애니메이션에서 attack trigger 값이 4
             _animator.SetInteger(AnimatorHash.TriggerNumber, context.TriggerNumber);
